requirements.txt
fastapi
uvicorn
python-dotenv
deepgram-sdk
websockets
pyaudio
numpy
webrtcvad
faster-whisper
requests
soundfile
google-genai

